# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Radar Vecinal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /panic-alerts:
    get:
      operationId: getPanicAlerts
      tags: [alerts]
      summary: Get all panic alerts
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: districtId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of panic alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/PanicAlert"
                required:
                  - alerts
    post:
      operationId: createPanicAlert
      tags: [alerts]
      summary: Create a panic alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePanicAlertInput"
      responses:
        "201":
          description: Panic alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanicAlert"
  /panic-alerts/stream:
    get:
      operationId: streamPanicAlerts
      tags: [alerts]
      summary: SSE stream of panic alerts
      parameters:
        - name: districtId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of panic alert events
          content:
            text/event-stream:
              schema:
                type: string
  /missing-persons:
    get:
      operationId: getMissingPersons
      tags: [missing]
      summary: Get missing person alerts
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: districtId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of missing person alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/MissingPerson"
                required:
                  - alerts
    post:
      operationId: createMissingPerson
      tags: [missing]
      summary: Create a missing person alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMissingPersonInput"
      responses:
        "201":
          description: Missing person alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingPerson"
  /missing-persons/{id}:
    patch:
      operationId: updateMissingPerson
      tags: [missing]
      summary: Update a missing person alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMissingPersonInput"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingPerson"
    delete:
      operationId: deleteMissingPerson
      tags: [missing]
      summary: Delete a missing person alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Alert deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  id:
                    type: string
                required:
                  - success
                  - id
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PanicAlert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - robbery
            - medical
            - fight
            - fire
            - missing_person
            - other
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        authorName:
          type: string
        sector:
          type: string
        districtId:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - latitude
        - longitude
        - authorName
        - sector
        - districtId
        - isActive
        - createdAt
    CreatePanicAlertInput:
      type: object
      properties:
        type:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        authorName:
          type: string
        sector:
          type: string
        districtId:
          type: integer
      required:
        - type
        - latitude
        - longitude
        - authorName
        - sector
    MissingPerson:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: number
          nullable: true
        clothing:
          type: string
        photoUrl:
          type: string
          nullable: true
        lastSeenLatitude:
          type: number
        lastSeenLongitude:
          type: number
        lastSeenAddress:
          type: string
        lastSeenAt:
          type: string
        contactInfo:
          type: string
        status:
          type: string
          enum:
            - active
            - found
            - archived
        reportedBy:
          type: string
        districtId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - clothing
        - lastSeenLatitude
        - lastSeenLongitude
        - lastSeenAddress
        - lastSeenAt
        - contactInfo
        - status
        - reportedBy
        - districtId
        - createdAt
    CreateMissingPersonInput:
      type: object
      properties:
        name:
          type: string
        age:
          type: number
          nullable: true
        clothing:
          type: string
        photoUrl:
          type: string
          nullable: true
        lastSeenLatitude:
          type: number
        lastSeenLongitude:
          type: number
        lastSeenAddress:
          type: string
        lastSeenAt:
          type: string
        contactInfo:
          type: string
        reportedBy:
          type: string
        districtId:
          type: integer
      required:
        - name
        - clothing
        - lastSeenLatitude
        - lastSeenLongitude
        - lastSeenAddress
        - lastSeenAt
        - contactInfo
        - reportedBy
    UpdateMissingPersonInput:
      type: object
      properties:
        status:
          type: string
        clothing:
          type: string
        photoUrl:
          type: string
          nullable: true
...