# lib/api-zod/src/generated/types/updateMissingPersonInput.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Api
 * Radar Vecinal API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMissingPersonInput {
  status?: string;
  clothing?: string;
  photoUrl?: string | null;
}